skills: move svgs into icon folder structure
