+++ icons/svgs/DockerSVG.tsx
