+++ icons/svgs/GatsbySVG.tsx
