+++ icons/svgs/NginxSVG.tsx
