+++ icons/svgs/NodeSVG.tsx
