+++ icons/svgs/ReactSVG.tsx
