+++ icons/svgs/TailwindSVG.tsx
