+++ icons/svgs/VueSVG.tsx
