+++ icons/svgs/VuetifySVG.tsx
